requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==5.0.1
django-import-export==3.3.6
django-jazzmin==2.6.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
magic-filter==1.0.11
MarkupPy==1.14
multidict==6.0.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
psycopg==3.1.17
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
urllib3==2.0.4
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
